refactor: 디자이너 요구사항 반영
