--- src/app/_main/components/slot-machine/machine-test.tsx
@@ -1,5 +1,0 @@
-import { PropsWithChildren } from "react"
-
-export default function MachineTest(props: PropsWithChildren) {
-  return <div className="phone-size bg-gray-950">{props.children}</div>
-}
