+++ src/hooks/useSlotMachine.tsx
@@ -14,6 +14,14 @@ type VariantProps = {
   filter?: string
 }
 
+/**
+ *
+ * @param textList: 랜덤으로 골라질 값들의 배열을 넣습니다.
+ * @param duration: 룰렛이 얼마나 오래 돌아갈지 결정합니다.
+ * @returns SlotMachine: 랜덤으로 돌아가는 문자만 들어가있는 컴포넌트.
+ * @returns randomStart: 랜덤으로 돌아가는걸 시작하는 함수
+ * @returns text: 랜덤으로 선택된 문자.
+ */
 export default function useSlotMachine({ textList, duration }: Props) {
   const [count, setCount] = useState(Math.floor(Math.random() * textList.length))
   const [currentIndex, setCurrentIndex] = useState(0)
